fix: 본인 서재 조회시, 프로필 사진이 변경중인지 확인하는 값 추가
