+++ src/main/java/com/bookbla/americano/domain/alarm/service/MemberTokenService.java
@@ -1,6 +1,6 @@
-package com.bookbla.americano.domain.member.service;
+package com.bookbla.americano.domain.alarm.service;
 
-import com.bookbla.americano.domain.member.controller.dto.request.MemberTokenCreateRequest;
+import com.bookbla.americano.domain.alarm.controller.dto.request.MemberTokenCreateRequest;
 import com.bookbla.americano.domain.member.controller.dto.response.MemberTokenCreateResponse;
 
 public interface MemberTokenService {
